tests/wpt/web-platform-tests/app-history/app-history-entry/entries-array-equality.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_not_equals(appHistory.entries(), appHistory.entries());
}, "appHistory.entries() should not return an identical object on repeated invocations");
</script>
